requirements.txt
asttokens==2.2.1
backcall==0.2.0
charset-normalizer==3.0.1
colorful==0.5.5
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
einops==0.6.0
entrypoints==0.4
executing==1.2.0
faiss-gpu==1.7.2
fonttools==4.38.0
greenlet==2.0.1
h5py==3.7.0
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-resources==5.10.2
ipdb==0.13.11
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter-client==7.4.9
jupyter-core==5.1.3
kiwisolver==1.4.4
kornia==0.6.9
llvmlite==0.39.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
more-itertools==9.0.0
msgpack==1.0.4
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.2
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.14.0
pynvim==0.4.3
pyod==1.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyyaml==6.0
pyzmq==25.0.0
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
statsmodels==0.13.5
tensorboardx==2.5.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing-extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
yachalk==0.1.5
zipp==3.11.0
